Fix : exportToCanvas() passing wrong prop to <Excalidraw />
In the example of exportToCanvas() usage, the prop ref is passed to the element. Ref can't be passed to function elements so it fires an error when using the example as is in a basic Excalidraw React app. The correct prop to pass in order to be able to get the elements needed by the exportToCanvas() method we need to retrieve the excalidrawAPI using the excalidrawAPI prop, as I corrected